Split shipping tests
